backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0.post2
python-multipart==0.0.6
hubspot-api-client==7.0.0
openai==1.3.0
pandas==2.1.0
numpy==1.24.3
chart.js==3.9.1
python-dotenv==1.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0.post2
python-multipart==0.0.6
